+++ example/lib/main.dart
@@ -33,14 +33,17 @@ class MultiSplitViewExample extends StatefulWidget {
 }
 
 class _MultiSplitViewExampleState extends State<MultiSplitViewExample> {
+  final int _max = 40;
   int _horizontalVisibleWidgets = 3;
+
+  MultiSplitViewController controller =
+      MultiSplitViewController(weights: [0.01]);
 
   @override
   Widget build(BuildContext context) {
-    int max = widget._colors.length;
     Widget buttons = Container(
         child: Row(children: [
-          Text('Horizontal widgets: $_horizontalVisibleWidgets / $max'),
+          Text('Horizontal widgets: $_horizontalVisibleWidgets / $_max'),
           SizedBox(width: 16),
           ElevatedButton(child: Text('Add'), onPressed: _onAddButtonClick),
           SizedBox(width: 16),
@@ -52,17 +55,15 @@ class _MultiSplitViewExampleState extends State<MultiSplitViewExample> {
     List<Widget> children = List.empty(growable: true);
     for (int i = 0; i < _horizontalVisibleWidgets; i++) {
       Widget view = Container(
-        width: 300,
-        height: 300,
         child: Center(child: Text("View " + (i + 1).toString())),
-        color: widget._colors[i],
+        color: widget._colors[i % widget._colors.length],
       );
       children.add(view);
     }
 
     MultiSplitView multiSplitView = MultiSplitView(
         children: children,
-        controller: MultiSplitViewController(weights: [0.1]),
+        controller: controller,
         onSizeChange: _onSizeChange);
 
     return Scaffold(
@@ -84,8 +85,7 @@ class _MultiSplitViewExampleState extends State<MultiSplitViewExample> {
 
   _onAddButtonClick() {
     setState(() {
-      _horizontalVisibleWidgets =
-          min(widget._colors.length, _horizontalVisibleWidgets + 1);
+      _horizontalVisibleWidgets = min(_max, _horizontalVisibleWidgets + 1);
     });
   }
 }